requirements.txt
Django==1.6.5
DjangoRestless==0.0.9
MySQL-python==1.2.5
argparse==1.2.1
distribute==0.6.24
gunicorn==19.0.0
six==1.7.3
wsgiref==0.1.2
